Update README.md
In the project "Sentiment Analysis Using Machine Learning and NLP," we delve into the fascinating realm of Natural Language Processing (NLP) and Machine Learning to unravel the emotions embedded in textual data. By employing advanced algorithms and techniques, we decipher the sentiments—be they positive, negative, or neutral—conveyed within the text. Our comprehensive approach not only involves the technical prowess of machine learning but also harnesses the power of language understanding. Join us on this journey as we extract and interpret sentiments, enabling us to gain valuable insights from the vast sea of textual information that surrounds us.